increase gap on hero-box to fix the layout
Basically hero box was not taking 100% parent width then the page is in full screen
leaving 50px on the right of the parent.
The child are spaced-between without any margin. The parent of hero-box has 300px padding
on x-axis but that isn't cousing the issue.
I have done the same thing to header and it's working properly.